requirements.txt
pyTelegramBotAPI==4.30.0
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3
pymongo==4.6.0